Send appSettings in client info event
Assurance Andorid SDK currently does not send "appSettings" as part of
the client info event. This is becuase of an earlier limitation in the
services in handling large payloads for this specific event.
Assurance services have now been enhanced to handle large payloads, so it
is safe to send the appSettings now.