Prevent outbound buffer overwrite during chunking
For large Assurance events, the way the outbound event  is being read
and converted to a string can cause un-necessary content being appended to
the last chunk. This is because the content is read in to a buffer and
converted to a string directly irrespective of the acutal bytes read.
So if the last chunk of data is smaller than the buffer size the remaining
part from the buffer could be included.

As a solution, modify the reading algorithm to store the number of bytes
actually read. Then when creating the string use only the portion of the
bytes actually read.