Terminate jobs when quick connect is cancelled
Currently, the device status check job is not being cancelled
when the 'Cancel' button is clicked before the QuickConnect
status check succeeds. This was a miss when porting from 2.x.
While this is a rare case, it is still possible to Cancel the
flow when waiting.

As a solution invoke QuickConnectManager.cancel() when Cancel is
clicked.