Disable content and file access from WebView
Assurance SDK does not need the WebView to allow a file scheme URL
to access other file scheme URLs. We do not `setAllowFileAccess`
value explicitly to false. For applications targeting API 29 and
below this defaults to true if not explicitly set. So set this flag
to false to tighten access.

Assurance SDK does not need to access WebView to access content data
on the device. We do not explicitly set the value to false. Set this
flag to false to tighten acces.