MOB-22620: Fix concurrent modification exception during session close
Problem:

`notifyTerminationAndRemoveStatusListeners()` iterates over all the connection
status listeners to notify disconnection and un-registers them. Since the
de-registration process involves removing the listeners from the set, iterating over
the same list for removal leads to ConcurrentModificationException and not all listeners
are notified of the error. This results in the connecting icon spinning forever.

Solution:

Iterate for  de-registration over a copy of the listeners to prevent the exception
and propogate the error.